[20190107] Update SegTree

+++ 트리/SegTree.cpp
@@ -1,39 +1,39 @@
 struct SegTree{
   int n;
-  vector<int> sum;
-  vector<int> lazy;
-  SegTree(const vector<int>& a){
+  vector<ll> sum;
+  vector<ll> lazy;
+  SegTree(const vector<ll>& a){
     n = a.size();
     sum.resize(n * 4);
     lazy.resize(n * 4);
     init(a, 0, n-1, 1);
   }
-  int init(const vector<int>& a, int l, int r, int node){
+  ll init(const vector<ll>& a, int l, int r, int node){
     if(l == r) return sum[node] = a[l];
     int mid = (l + r) / 2;
-    int lSum = init(a, l, mid, node*2);
-    int rSum = init(a, mid+1, r, node*2+1);
+    ll lSum = init(a, l, mid, node*2);
+    ll rSum = init(a, mid+1, r, node*2+1);
     return sum[node] = lSum + rSum;
   }
-  int query(int l, int r, int node, int nl, int nr){
+  ll query(int l, int r, int node, int nl, int nr){
     update_lazy(nl,nr,node);
     if(r < nl || nr < l) return 0;
     if(l <= nl && nr <= r) return sum[node];
     int mid = (nl + nr) / 2;
     return query(l, r, node*2, nl, mid) + query(l, r, node*2+1, mid+1, nr);
   }
-  int query(int l, int r){
+  ll query(int l, int r){
     return query(l, r, 1, 0, n-1);
   }
   //array[index]=newValue로 바뀌었을 때 node를 루트로 하는 구간트리를 갱신
-  int update(int index, int newValue, int node, int nl, int nr){
+  ll update(int index, ll newValue, int node, int nl, int nr){
     if(index < nl || nr < index) return sum[node];
     if(nl == nr) return sum[node] = newValue;
     int mid = (nl + nr) / 2;
     return sum[node] = update(index, newValue, node*2, nl, mid) +
               update(index, newValue, node*2+1, mid+1, nr);
   }
-  int update(int index, int newValue){
+  ll update(int index, ll newValue){
     return update(index, newValue, 1, 0, n-1);
   }
   void update_lazy(int l, int r, int node){
@@ -47,7 +47,7 @@ struct SegTree{
     }
   }
   // left~right구간을 diff만큼 증가하였을 때 node를 루트로 하는 구간트리를 갱신
-  void update_range(int l, int r, int diff, int node, int nl, int nr){
+  void update_range(int l, int r, ll diff, int node, int nl, int nr){
     update_lazy(nl,nr,node);
     if(r < nl || nr < l) return;
     if(l <= nl && nr <= r){
@@ -63,7 +63,7 @@ struct SegTree{
     update_range(l, r, diff, node*2+1, mid+1, nr);
     sum[node] = sum[node*2] + sum[node*2+1];
   }
-  void update_range(int l, int r, int diff){
+  void update_range(int l, int r, ll diff){
     update_range(l, r, diff, 1, 0, n-1);
   }
 };
